refactor: friend 도메인 구조 변경
