+++ build.gradle
@@ -103,31 +103,31 @@ jacocoTestReport {
 
 // Per-domain coverage reports
 def domains = [
-    'auth', 'chat', 'image', 'order', 'payment',
-    'product', 'search', 'seller', 'shipping', 'store', 'term', 'user'
+        'auth', 'chat', 'friend', 'image', 'order', 'payment',
+        'product', 'search', 'seller', 'shipping', 'store', 'term', 'user'
 ]
 
 domains.each { domain ->
     tasks.register("jacocoReport${domain.capitalize()}", JacocoReport) {
-        group       = 'verification'
+        group = 'verification'
         description = "JaCoCo coverage report for domain: ${domain}"
         dependsOn tasks.named('test')
         executionData fileTree(layout.buildDirectory.dir("jacoco")).include("*.exec")
         sourceDirectories.setFrom(sourceSets.main.java.srcDirs)
         classDirectories.setFrom(
-            sourceSets.main.output.classesDirs.asFileTree.matching {
-                include "**/domain/${domain}/**"
-            }
+                sourceSets.main.output.classesDirs.asFileTree.matching {
+                    include "**/domain/${domain}/**"
+                }
         )
         reports {
-            xml.required  = true
+            xml.required = true
             html.required = true
         }
     }
 }
 
 tasks.register('jacocoAllDomainReports') {
-    group       = 'verification'
+    group = 'verification'
     description = 'Generates JaCoCo coverage reports for all domains'
     dependsOn domains.collect { "jacocoReport${it.capitalize()}" }
 }
