+++ src/main/java/com/example/WonkaoTalk/domain/friend/dto/FriendAddRequest.java
@@ -1,4 +1,4 @@
-package com.example.WonkaoTalk.domain.user.friend.dto;
+package com.example.WonkaoTalk.domain.friend.dto;
 
 import jakarta.validation.constraints.NotNull;
