+++ src/main/java/com/example/WonkaoTalk/domain/friend/dto/FriendAddResponse.java
@@ -1,4 +1,4 @@
-package com.example.WonkaoTalk.domain.user.friend.dto;
+package com.example.WonkaoTalk.domain.friend.dto;
 
 import lombok.Builder;
