+++ src/main/java/com/example/WonkaoTalk/domain/friend/enums/FriendStatus.java
@@ -1,4 +1,4 @@
-package com.example.WonkaoTalk.domain.user.friend.enums;
+package com.example.WonkaoTalk.domain.friend.enums;
 
 public enum FriendStatus {
   ACTIVE,
